Fix links for dotted instructions
I noticed that clicking on instructions with `.` in their name (e.g. `abs.w`) didn't go to the section for the instruction, while things worked fine for instructions without `.` in their name. Looking at the HTML page source of the GitHub-rendered markdown, it looks like, somehow, the HTML fragments for the instruction sections lose the `.`, e.g. the fragment name for `abs.w` is not `abs.w` but `absw`. So I removed the dot from the fragments references like `(#absw)` and the link appear to work in the preview.